+++ testing.go
@@ -9,10 +9,10 @@ import (
 	"io"
 	"net"
 	"net/rpc"
+	"testing"
 
 	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/go-plugin/internal/plugin"
-	"github.com/mitchellh/go-testing-interface"
+	"github.com/hashicorp/go-plugin/internal/grpcmux"
 	"google.golang.org/grpc"
 )
 
@@ -33,7 +33,7 @@ type TestOptions struct {
 
 // TestConn is a helper function for returning a client and server
 // net.Conn connected to each other.
-func TestConn(t testing.T) (net.Conn, net.Conn) {
+func TestConn(t testing.TB) (net.Conn, net.Conn) {
 	// Listen to any local port. This listener will be closed
 	// after a single connection is established.
 	l, err := net.Listen("tcp", "127.0.0.1:0")
@@ -67,7 +67,7 @@ func TestConn(t testing.T) (net.Conn, net.Conn) {
 }
 
 // TestRPCConn returns a rpc client and server connected to each other.
-func TestRPCConn(t testing.T) (*rpc.Client, *rpc.Server) {
+func TestRPCConn(t testing.TB) (*rpc.Client, *rpc.Server) {
 	clientConn, serverConn := TestConn(t)
 
 	server := rpc.NewServer()
@@ -79,7 +79,7 @@ func TestRPCConn(t testing.T) (*rpc.Client, *rpc.Server) {
 
 // TestPluginRPCConn returns a plugin RPC client and server that are connected
 // together and configured.
-func TestPluginRPCConn(t testing.T, ps map[string]Plugin, opts *TestOptions) (*RPCClient, *RPCServer) {
+func TestPluginRPCConn(t testing.TB, ps map[string]Plugin, opts *TestOptions) (*RPCClient, *RPCServer) {
 	// Create two net.Conns we can use to shuttle our control connection
 	clientConn, serverConn := TestConn(t)
 
@@ -107,7 +107,7 @@ func TestPluginRPCConn(t testing.T, ps map[string]Plugin, opts *TestOptions) (*R
 // TestGRPCConn returns a gRPC client conn and grpc server that are connected
 // together and configured. The register function is used to register services
 // prior to the Serve call. This is used to test gRPC connections.
-func TestGRPCConn(t testing.T, register func(*grpc.Server)) (*grpc.ClientConn, *grpc.Server) {
+func TestGRPCConn(t testing.TB, register func(*grpc.Server)) (*grpc.ClientConn, *grpc.Server) {
 	// Create a listener
 	l, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
@@ -135,49 +135,51 @@ func TestGRPCConn(t testing.T, register func(*grpc.Server)) (*grpc.ClientConn, *
 
 // TestPluginGRPCConn returns a plugin gRPC client and server that are connected
 // together and configured. This is used to test gRPC connections.
-func TestPluginGRPCConn(t testing.T, ps map[string]Plugin) (*GRPCClient, *GRPCServer) {
+func TestPluginGRPCConn(t testing.TB, multiplex bool, ps map[string]Plugin) (*GRPCClient, *GRPCServer) {
 	// Create a listener
-	l, err := net.Listen("tcp", "127.0.0.1:0")
+	ln, err := serverListener(UnixSocketConfig{})
 	if err != nil {
-		t.Fatalf("err: %s", err)
+		t.Fatal(err)
 	}
 
+	logger := hclog.New(&hclog.LoggerOptions{
+		Level: hclog.Debug,
+	})
+
 	// Start up the server
+	var muxer *grpcmux.GRPCServerMuxer
+	if multiplex {
+		muxer = grpcmux.NewGRPCServerMuxer(logger, ln)
+		ln = muxer
+	}
 	server := &GRPCServer{
 		Plugins: ps,
 		DoneCh:  make(chan struct{}),
 		Server:  DefaultGRPCServer,
 		Stdout:  new(bytes.Buffer),
 		Stderr:  new(bytes.Buffer),
-		logger:  hclog.Default(),
+		logger:  logger,
+		muxer:   muxer,
 	}
 	if err := server.Init(); err != nil {
 		t.Fatalf("err: %s", err)
 	}
-	go server.Serve(l)
+	go server.Serve(ln)
 
-	// Connect to the server
-	conn, err := grpc.Dial(
-		l.Addr().String(),
-		grpc.WithBlock(),
-		grpc.WithInsecure())
+	client := &Client{
+		address:  ln.Addr(),
+		protocol: ProtocolGRPC,
+		config: &ClientConfig{
+			Plugins:             ps,
+			GRPCBrokerMultiplex: multiplex,
+		},
+		logger: logger,
+	}
+
+	grpcClient, err := newGRPCClient(context.Background(), client)
 	if err != nil {
-		t.Fatalf("err: %s", err)
+		t.Fatal(err)
 	}
 
-	brokerGRPCClient := newGRPCBrokerClient(conn)
-	broker := newGRPCBroker(brokerGRPCClient, nil, UnixSocketConfig{}, nil)
-	go broker.Run()
-	go brokerGRPCClient.StartStream()
-
-	// Create the client
-	client := &GRPCClient{
-		Conn:       conn,
-		Plugins:    ps,
-		broker:     broker,
-		doneCtx:    context.Background(),
-		controller: plugin.NewGRPCControllerClient(conn),
-	}
-
-	return client, server
+	return grpcClient, server
 }